Assignment1/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
google-cloud-storage==1.41.1
Werkzeug